requirements.txt
# Core dependencies
tweepy>=4.14.0  # Twitter API client
openai>=1.60.1  # OpenAI GPT integration
python-dotenv>=1.0.0  # Environment variable management

# Development dependencies
pytest>=7.4.0   # Testing framework
flake8>=7.1.1   # Code linting

# Installation note:
# If you encounter SSL certificate errors, use:
# pip install --trusted-host pypi.org --trusted-host files.pythonhosted.org -r requirements.txt
